# Performance: store path in stack to avoid copying context