Fixes #325 

This PR is created to dramatically improve performance for go-yaml library especially when parsing larger YAML files.

This is the methodology that I used to arrive at the current solution:
1. I use `go test` tool to create a CPU profile when doing `parser.ParseBytes`.
2. When I analyzed the CPUProfile, I found that `withChild` and `withIndex` are the two functions that dominated the CPU utilization.
3. Reading the source code, I found that both functions call `parser.copy`. This operation is performed repeatedly when parsing the YAML tokens.
4. Taking a deeper look, the essence of this copy operation is to pass the correct node path when parsing child node.
5. We can achieve the same result by managing a stack. We can push when parsing child node and pop after we finish.
6. This way, we improve performance by avoiding excessive memory allocation caused by copying the `context.tokens`.

I realized the implementation can be refactored further to adjust the way of using context, but I aim for minimum change in this PR. Please let me know if you have any suggestions 🙏 . Thank you!

Checklist:
- [x] Describe the purpose for which you created this PR.
- [ ] Create test code that corresponds to the modification