@fzipi 2024-04-07
Looks good to us also. See benchmarks in the related issue.