@goccy 2024-10-31
@yoelsusanto Thank you for your contribution. Also, Sorry for the late reply.
I'd like to improve the process of `pop` from the stack, so I created a PR on my side.